refactor(cli): extract samples sub-app to cli_samples.py (ADR-059 V9)
